+++ sys/ufs/ufs/ufs_quota.c
@@ -283,8 +283,8 @@ chkdqchg(struct inode *ip, long change, struct ucred *cred, int type)
 	 */
 	if (ncurblocks >= dq->dq_bsoftlimit && dq->dq_bsoftlimit) {
 		if (dq->dq_curblocks < dq->dq_bsoftlimit) {
-			dq->dq_btime = time_second +
-			    ip->i_ump->um_btime[type];
+			dq->dq_btime = (uint32_t)(time_second +
+			    ip->i_ump->um_btime[type]);
 			if (DIP(ip, uid) == cred->cr_uid)
 				uprintf("\n%s: warning, %s %s\n",
 				    ITOV(ip)->v_mount->mnt_stat.f_mntonname,
@@ -404,8 +404,8 @@ chkiqchg(struct inode *ip, long change, struct ucred *cred, int type)
 	 */
 	if (ncurinodes >= dq->dq_isoftlimit && dq->dq_isoftlimit) {
 		if (dq->dq_curinodes < dq->dq_isoftlimit) {
-			dq->dq_itime = time_second +
-			    ip->i_ump->um_itime[type];
+			dq->dq_itime = (uint32_t)(time_second +
+			    ip->i_ump->um_itime[type]);
 			if (DIP(ip, uid) == cred->cr_uid)
 				uprintf("\n%s: warning, %s %s\n",
 				    ITOV(ip)->v_mount->mnt_stat.f_mntonname,
@@ -685,11 +685,13 @@ setquota(struct mount *mp, u_long id, int type, caddr_t addr)
 	if (newlim.dqb_bsoftlimit &&
 	    dq->dq_curblocks >= newlim.dqb_bsoftlimit &&
 	    (dq->dq_bsoftlimit == 0 || dq->dq_curblocks < dq->dq_bsoftlimit))
-		newlim.dqb_btime = time_second + ump->um_btime[type];
+		newlim.dqb_btime = (uint32_t)(time_second +
+		    ump->um_btime[type]);
 	if (newlim.dqb_isoftlimit &&
 	    dq->dq_curinodes >= newlim.dqb_isoftlimit &&
 	    (dq->dq_isoftlimit == 0 || dq->dq_curinodes < dq->dq_isoftlimit))
-		newlim.dqb_itime = time_second + ump->um_itime[type];
+		newlim.dqb_itime = (uint32_t)(time_second +
+		    ump->um_itime[type]);
 	dq->dq_dqb = newlim;
 	if (dq->dq_curblocks < dq->dq_bsoftlimit)
 		dq->dq_flags &= ~DQ_BLKS;
@@ -741,10 +743,10 @@ setuse(struct mount *mp, u_long id, int type, caddr_t addr)
 	 */
 	if (dq->dq_bsoftlimit && dq->dq_curblocks < dq->dq_bsoftlimit &&
 	    usage.dqb_curblocks >= dq->dq_bsoftlimit)
-		dq->dq_btime = time_second + ump->um_btime[type];
+		dq->dq_btime = (uint32_t)(time_second + ump->um_btime[type]);
 	if (dq->dq_isoftlimit && dq->dq_curinodes < dq->dq_isoftlimit &&
 	    usage.dqb_curinodes >= dq->dq_isoftlimit)
-		dq->dq_itime = time_second + ump->um_itime[type];
+		dq->dq_itime = (uint32_t)(time_second + ump->um_itime[type]);
 	dq->dq_curblocks = usage.dqb_curblocks;
 	dq->dq_curinodes = usage.dqb_curinodes;
 	if (dq->dq_curblocks < dq->dq_bsoftlimit)
@@ -903,7 +905,7 @@ dqget(struct vnode *vp, u_long id, struct ufsmount *ump, int type,
 	LIST_INSERT_HEAD(dqh, dq, dq_hash);
 	dqref(dq);
 	dq->dq_flags = DQ_LOCK;
-	dq->dq_id = id;
+	dq->dq_id = (uint32_t)id;
 	dq->dq_vp = dqvp;
 	dq->dq_type = type;
 	crhold(ump->um_cred[type]);
@@ -944,9 +946,11 @@ dqget(struct vnode *vp, u_long id, struct ufsmount *ump, int type,
 		dq->dq_flags |= DQ_FAKE;
 	if (dq->dq_id != 0) {
 		if (dq->dq_btime == 0)
-			dq->dq_btime = time_second + ump->um_btime[type];
+			dq->dq_btime = (uint32_t)(time_second +
+			    ump->um_btime[type]);
 		if (dq->dq_itime == 0)
-			dq->dq_itime = time_second + ump->um_itime[type];
+			dq->dq_itime = (uint32_t)(time_second +
+			    ump->um_itime[type]);
 	}
 	*dqp = dq;
 	return (0);
